Add ConnectFourGUI file with class and main
Create init method for ConnectFourGUI class that starts the game
4-to-Connect. This class also has methods to change to and from the
start, help, play and exit screen. The main function creates an instance
of the ConnectFourGUI and exits the program when requested by a user.

The GUI and its instantiation for the Connect-Four-Project is necessary
in order for the user to interact with the game 4-to-Connect. The file
ConnectFourGUI relies on the implementation of the ConnectFourBoard
class.