Implement Main class working with user input
Implemented Main class contains all classes as its attributes. The
main function can return the winner of the game as a string using the
methods in Board class. So that adding the function to the main
function should properly run the game without GUI. Yet, the main class
can not update the GUI properly hence inserting GUI methods to
appropriate method is required.